requirements.txt
Django==5.2.6
djangorestframework==3.16.0
djangorestframework-simplejwt
django-cors-headers
django-filter==25.1
Pillow